+++ src/main/java/at/hannibal2/skyhanni/config/features/MiscConfig.java
@@ -535,6 +535,31 @@ public class MiscConfig {
 
     }
 
+
+    @Expose
+    @ConfigOption(name = "Highlight Party Members", desc = "")
+    @Accordion
+    public HighlightPartyMembers highlightPartyMembers = new HighlightPartyMembers();
+
+    public static class HighlightPartyMembers {
+
+        @Expose
+        @ConfigOption(name = "Enabled", desc = "Marking party members with a bright outline to better find them in the world.")
+        @ConfigEditorBoolean
+        @FeatureToggle
+        public boolean enabled = false;
+
+        @Expose
+        @ConfigOption(
+                name = "Outline Color",
+                desc = "The color to outline party members in."
+        )
+        @ConfigEditorColour
+        public String outlineColor = "0:245:85:255:85";
+
+    }
+
+
     @Expose
     @ConfigOption(name = "Compact Tab List", desc = "")
     @Accordion
@@ -611,7 +636,64 @@ public class MiscConfig {
             @ConfigOption(name = "Mark Special Persons", desc = "Show speical icons behind the name of guild members, party members, friends, and marked players.")
             @ConfigEditorBoolean
             public boolean markSpecialPersons = false;
+
+            @Expose
+            @ConfigOption(name = "Mark SkyHanni Devs", desc = "Adds a §c:O §7behind the tablist name of SkyHanni's creators.")
+            @ConfigEditorBoolean
+            public boolean markSkyHanniDevs = false;
         }
+    }
+
+    @Expose
+    @ConfigOption(name = "Kick Duration", desc = "")
+    @Accordion
+    public KickDurationConfig kickDuration = new KickDurationConfig();
+
+    public static class KickDurationConfig {
+
+        @Expose
+        @ConfigOption(
+                name = "Enabled",
+                desc = "Show in the Hypixel lobby since when you were last kicked from SkyBlock (" +
+                        "useful if you get blocked because of '§cYou were kicked while joining that server!§7')."
+        )
+        @ConfigEditorBoolean
+        @FeatureToggle
+        public boolean enabled = true;
+
+        @Expose
+        @ConfigOption(name = "Warn Time", desc = "Send warning and sound this seconds after a SkyBlock kick.")
+        @ConfigEditorSlider(
+                minValue = 5,
+                maxValue = 300,
+                minStep = 1
+        )
+        public Property<Integer> warnTime = Property.of(60);
+
+        @Expose
+        public Position position = new Position(400, 200, 1.3f);
+    }
+
+    @Expose
+    @ConfigOption(name = "Player Chat Symbols", desc = "")
+    @Accordion
+    public ChatSymbols chatSymbols = new ChatSymbols();
+
+    public static class ChatSymbols {
+
+        @Expose
+        @ConfigOption(name = "Enabled", desc = "Adds extra symbols to the chat such as those from ironman, " +
+                "stranded, bingo or nether factions and places them next to your regular player emblems. " +
+                "Also allows emblems without sb level in chat.")
+        @ConfigEditorBoolean
+        @FeatureToggle
+        public boolean enabled = true;
+
+        @Expose
+        @ConfigOption(name = "Chat Symbol Location", desc = "Determines where the symbols should go in chat in relation to the " +
+                "player's name. Hidden will hide all emblems from the chat. §eRequires above setting to be on to hide the symbols.")
+        @ConfigEditorDropdown(values = {"Left", "Right", "Hidden"})
+        public int symbolLocation = 0;
     }
 
     @Expose
@@ -716,6 +798,17 @@ public class MiscConfig {
     @ConfigEditorBoolean
     @FeatureToggle
     public boolean fixNeuHeavyPearls = true;
+
+    @Expose
+    @ConfigOption(
+            name = "Time In Limbo",
+            desc = "Show the time since you entered the limbo.")
+    @ConfigEditorBoolean
+    @FeatureToggle
+    public boolean showTimeInLimbo = true;
+
+    @Expose
+    public Position showTimeInLimboPosition = new Position(400, 200, 1.3f);
 
     @Expose
     public Position inventoryLoadPos = new Position(394, 124, false, true);