+++ src/main/java/at/hannibal2/skyhanni/data/RenderData.kt
@@ -1,7 +1,9 @@
 package at.hannibal2.skyhanni.data
 
 import at.hannibal2.skyhanni.events.GuiRenderEvent
+import at.hannibal2.skyhanni.events.LorenzRenderWorldEvent
 import at.hannibal2.skyhanni.features.garden.fortuneguide.FFGuideGUI
+import at.hannibal2.skyhanni.test.SkyHanniDebugsAndTests
 import at.hannibal2.skyhanni.features.misc.visualwords.VisualWordGui
 import net.minecraft.client.Minecraft
 import net.minecraft.client.gui.inventory.GuiChest
@@ -9,13 +11,15 @@ import net.minecraft.client.gui.inventory.GuiInventory
 import net.minecraft.client.renderer.GlStateManager
 import net.minecraftforge.client.event.GuiScreenEvent
 import net.minecraftforge.client.event.RenderGameOverlayEvent
+import net.minecraftforge.client.event.RenderWorldLastEvent
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 
-class RenderGuiData {
+class RenderData {
 
     @SubscribeEvent
     fun onRenderOverlay(event: RenderGameOverlayEvent.Pre) {
         if (event.type != RenderGameOverlayEvent.ElementType.HOTBAR) return
+        if (!SkyHanniDebugsAndTests.globalRender) return
         if (GuiEditManager.isInGui() || FFGuideGUI.isInGui() || VisualWordGui.isInGui()) return
 
         GuiRenderEvent.GuiOverlayRenderEvent().postAndCatch()
@@ -23,6 +27,7 @@ class RenderGuiData {
 
     @SubscribeEvent
     fun onBackgroundDraw(event: GuiScreenEvent.BackgroundDrawnEvent) {
+        if (!SkyHanniDebugsAndTests.globalRender) return
         if (GuiEditManager.isInGui() || FFGuideGUI.isInGui() || VisualWordGui.isInGui()) return
         val currentScreen = Minecraft.getMinecraft().currentScreen ?: return
         if (currentScreen !is GuiInventory && currentScreen !is GuiChest) return
@@ -37,5 +42,11 @@ class RenderGuiData {
         GuiRenderEvent.ChestGuiOverlayRenderEvent().postAndCatch()
 
         GlStateManager.popMatrix()
+    }
+
+    @SubscribeEvent
+    fun onRenderWorld(event: RenderWorldLastEvent) {
+        if (!SkyHanniDebugsAndTests.globalRender) return
+        LorenzRenderWorldEvent(event.partialTicks).postAndCatch()
     }
 }