+++ src/main/java/at/hannibal2/skyhanni/features/inventory/QuickCraftFeatures.kt
@@ -7,6 +7,7 @@ import at.hannibal2.skyhanni.events.LorenzToolTipEvent
 import at.hannibal2.skyhanni.events.RepositoryReloadEvent
 import at.hannibal2.skyhanni.utils.InventoryUtils
 import at.hannibal2.skyhanni.utils.ItemUtils.name
+import at.hannibal2.skyhanni.utils.KeyboardManager
 import at.hannibal2.skyhanni.utils.LorenzColor
 import at.hannibal2.skyhanni.utils.LorenzUtils
 import at.hannibal2.skyhanni.utils.StringUtils.removeColor
@@ -38,15 +39,16 @@ class QuickCraftFeatures {
     @SubscribeEvent
     fun onBackgroundDrawn(event: GuiContainerEvent.BackgroundDrawnEvent) {
         if (!isEnabled()) return
-        if (LorenzUtils.isControlKeyDown()) return
+        if (KeyboardManager.isControlKeyDown()) return
         if (event.gui !is GuiChest) return
         val chest = event.gui.inventorySlots as ContainerChest
 
         for (slot in chest.inventorySlots) {
             if (slot == null) continue
             if (slot.slotNumber !in quickCraftSlots) continue
-            val stack = slot.stack
-            if (stack.name == "§cQuick Crafting Slot") continue
+            val stack = slot.stack ?: continue
+            val name = stack.name ?: continue
+            if (name == "§cQuick Crafting Slot") continue
             if (needsQuickCraftConfirmation(stack)) {
                 val color = LorenzColor.DARK_GRAY.addOpacity(180)
                 stack.background = color.rgb
@@ -60,7 +62,7 @@ class QuickCraftFeatures {
 
         val clickedItem = event.slot?.stack ?: return
 
-        if (!LorenzUtils.isControlKeyDown() && needsQuickCraftConfirmation(clickedItem)) {
+        if (!KeyboardManager.isControlKeyDown() && needsQuickCraftConfirmation(clickedItem)) {
             event.isCanceled = true
         }
     }