+++ src/main/java/at/hannibal2/skyhanni/features/inventory/RngMeterInventory.kt
@@ -46,7 +46,7 @@ class RngMeterInventory {
         if (SkyHanniMod.feature.inventory.rngMeterSelectedDrop && chestName.endsWith(" RNG Meter")) {
             for (slot in InventoryUtils.getItemsInOpenChest()) {
                 val stack = slot.stack
-                if (stack.getLore().any { it.contains("§aSELECTED") }) {
+                if (stack.getLore().any { it.contains("§a§lSELECTED") }) {
                     slot highlight LorenzColor.YELLOW
                 }
             }