+++ src/main/java/at/hannibal2/skyhanni/features/misc/PetExpTooltip.kt
@@ -3,8 +3,10 @@ package at.hannibal2.skyhanni.features.misc
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.utils.ItemUtils.getItemRarityOrNull
 import at.hannibal2.skyhanni.utils.ItemUtils.name
+import at.hannibal2.skyhanni.utils.KeyboardManager
 import at.hannibal2.skyhanni.utils.LorenzRarity
 import at.hannibal2.skyhanni.utils.LorenzUtils
+import at.hannibal2.skyhanni.utils.LorenzUtils.makeAccessible
 import at.hannibal2.skyhanni.utils.LorenzUtils.round
 import at.hannibal2.skyhanni.utils.NumberUtil
 import at.hannibal2.skyhanni.utils.NumberUtil.addSeparators
@@ -25,7 +27,7 @@ class PetExpTooltip {
     fun onItemTooltipLow(event: ItemTooltipEvent) {
         if (!LorenzUtils.inSkyBlock) return
         if (!config.petDisplay) return
-        if (!LorenzUtils.isShiftKeyDown() && !config.showAlways) return
+        if (!KeyboardManager.isShiftKeyDown() && !config.showAlways) return
 
         val itemStack = event.itemStack ?: return
         val petExperience = itemStack.getPetExp()?.round(1) ?: return
@@ -58,11 +60,23 @@ class PetExpTooltip {
 
         index = toolTip.indexOfFirst { it.contains("Progress to Level") }
         if (index != -1) {
-            val offset = if (NotEnoughUpdates.INSTANCE.config.tooltipTweaks.petExtendExp) 4 else 3
+
+            val offset = if (isNeuExtendedExpEnabled) 4 else 3
             return index + offset
         }
 
         return null
+    }
+
+    private val isNeuExtendedExpEnabled get() = fieldPetExtendExp.get(objectNeuTooltipTweaks) as Boolean
+
+    private val objectNeuTooltipTweaks by lazy {
+        val field = NotEnoughUpdates.INSTANCE.config.javaClass.getDeclaredField("tooltipTweaks")
+        field.makeAccessible().get(NotEnoughUpdates.INSTANCE.config)
+    }
+
+    private val fieldPetExtendExp by lazy {
+        objectNeuTooltipTweaks.javaClass.getDeclaredField("petExtendExp").makeAccessible()
     }
 
     private fun getMaxValues(petName: String, petExperience: Double): Pair<Int, Int> {