+++ src/main/java/at/hannibal2/skyhanni/test/SkyHanniConfigSearchResetCommand.kt
@@ -82,13 +82,19 @@ object SkyHanniConfigSearchResetCommand {
             rawJson = readFromClipboard
         }
 
-        val root: Any = if (term.startsWith("config")) {
-            SkyHanniMod.feature
-        } else if (term.startsWith("playerSpecific")) {
-            ProfileStorageData.playerSpecific ?: return "§cplayerSpecific is null!"
-        } else if (term.startsWith("profileSpecific")) {
-            ProfileStorageData.profileSpecific ?: return "§cprofileSpecific is null!"
-        } else return "§cUnknown config location!"
+        val root: Any = when {
+            term.startsWith("config") -> SkyHanniMod.feature
+
+            term.startsWith("playerSpecific") -> {
+                ProfileStorageData.playerSpecific ?: return "§cplayerSpecific is null!"
+            }
+
+            term.startsWith("profileSpecific") -> {
+                ProfileStorageData.profileSpecific ?: return "§cprofileSpecific is null!"
+            }
+
+            else -> return "§cUnknown config location!"
+        }
 
         val affectedElements = findConfigElements({ it.startsWith("$term.") }, { true }).size
         if (affectedElements > 3 && !args.contentEquals(lastCommand)) {